requirements.txt
apache-airflow
pandas
boto3
botocore
newsapi-python
dotenv
apache-airflow[amazon]

